fix: increase Typesense max-per-page server option
The forum sets a constant search comment count of 1000
( https://github.com/openedx/forum/blob/51ec507/forum/constants.py ),
and the Elasticsearch and Meilisearch backends respect that.
So to use this value for Typesense too,
we need to increase the max-per-page server option.
The default is 250 in Typesense.

Private-ref: https://tasks.opencraft.com/browse/BB-9975